INamedTypeSymbol 상속 검사 관련 함수 Extension으로 빼고 사용처에 적용
